withVectorStore_Implementation/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf
huggingface-hub
sentence-transformers
cassio
datasets
langchain
openai
tiktoken
pymupdf
langchain-community
cassandra-driver
streamlit
langchain-groq
langchain-chroma
langchain-huggingface